server/requirements.txt
SQLAlchemy>=2.0,<3.0
psycopg[binary]>=3.2        # PG driver (v3); alternative: psycopg2-binary
python-dotenv>=1.0
bcrypt>=4.1
